+++ commands/uim/commands/classes/schemacaches/clear.d
@@ -21,14 +21,14 @@ class DSchemacacheClearCommand : DCommand {
     return "schema_cache-clear";
   }
 
-  override override size_t execute(Json[string] arguments, IConsole aConsole = null) {
+  override override ulong execute(Json[string] arguments, IConsole aConsole = null) {
     return super.execute(arguments, aConsoleIo);
   }
 
   /* 
 
   // Display all routes in an application
-  override size_t execute(Json[string] arguments, IConsole aConsole = null) {
+  override ulong execute(Json[string] arguments, IConsole aConsole = null) {
     try {
       aConnection = ConnectionManager . get(to!string(commandArguments.getOption("connection")));
       assert(cast(Connection)aConnection);
