+++ oop/uim/oop/commands/interfaces.d
@@ -4,5 +4,5 @@ import uim.oop;
 @safe:
 
 interface ICommand : INamed {
-    size_t execute(Json[string] options, IConsole console = null);
+    ulong execute(Json[string] options, IConsole console = null);
 }
